# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "cell-analysis-backend",
        "description": "Backend API for cancer cell analysis application, apoptosis and viable cell",
        "contact": {
            "name": "Your Name",
            "email": "your.email@example.com"
        },
        "license": {
            "name": ""
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/images/{image_id}/analysis-history": {
            "get": {
                "tags": [
                    "AI Analysis"
                ],
                "summary": "Get analysis history for an image",
                "operationId": "get_analysis_history",
                "parameters": [
                    {
                        "name": "image_id",
                        "in": "path",
                        "description": "Image ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analysis history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_ImageAnalysisHistoryResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Image not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/images/{image_id}/analyze": {
            "post": {
                "tags": [
                    "AI Analysis"
                ],
                "summary": "Submit an image for AI analysis via RabbitMQ",
                "operationId": "analyze_image",
                "parameters": [
                    {
                        "name": "image_id",
                        "in": "path",
                        "description": "Image ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnalyzeImageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Analysis job created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_AnalyzeImageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Image not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
...
        "/api/v1/jobs/{job_id}": {
            "get": {
                "tags": [
                    "AI Analysis"
                ],
                "summary": "Get the status of an analysis job",
                "operationId": "get_job_status",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_JobStatusResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/jobs/{job_id}/result": {
            "get": {
                "tags": [
                    "AI Analysis"
                ],
                "summary": "Get the result of a completed analysis job",
                "operationId": "get_job_result",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analysis result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_AnalysisResultResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Result not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AnalysisHistorySummary": {
                "type": "object",
                "description": "Summary of a single analysis in history",
                "required": [
                    "job_id",
                    "status"
                ],
                "properties": {
                    "ai_model_version": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "avg_confidence_score": {
                        "type": [
                            "number",
                            "null"
                        ],
                        "format": "double"
                    },
                    "counts": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/CellCounts"
                            }
                        ]
                    },
                    "finished_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "job_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
...
            "AnalyzeImageRequest": {
                "type": "object",
                "description": "Request to analyze an image",
                "properties": {
                    "model_version": {
                        "type": "string",
                        "description": "AI model version to use (optional, defaults to latest)"
                    }
                }
            },
...
            "ApiError": {
                "type": "object",
                "description": "API error structure",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "ApiResponse_AnalysisResultResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Analysis result response",
                        "required": [
                            "result_id",
                            "job_id",
                            "image_id",
                            "counts",
                            "total_cells",
                            "avg_confidence_score",
                            "percentages",
                            "analyzed_at"
                        ],
                        "properties": {
                            "analyzed_at": {
                                "type": "string"
                            },
                            "avg_confidence_score": {
                                "type": "number",
                                "format": "double"
                            },
                            "counts": {
                                "$ref": "#/components/schemas/CellCounts"
                            },
                            "image_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "job_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "percentages": {
                                "$ref": "#/components/schemas/CellPercentages"
                            },
                            "raw_data": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/RawDetectionData"
                                    }
                                ]
                            },
                            "result_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "summary_data": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "total_cells": {
                                "type": "integer",
                                "format": "int32"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
            "ApiResponse_AnalyzeImageResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Response when submitting image for analysis",
                        "required": [
                            "job_id",
                            "image_id",
                            "status",
                            "ai_model_version",
                            "status_url",
                            "created_at"
                        ],
                        "properties": {
                            "ai_model_version": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "image_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "job_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "status": {
                                "type": "string"
                            },
                            "status_url": {
                                "type": "string"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "ApiResponse_ImageAnalysisHistoryResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Analysis history response for an image",
                        "required": [
                            "image_id",
                            "analyses",
                            "total"
                        ],
                        "properties": {
                            "analyses": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/AnalysisHistorySummary"
                                }
                            },
                            "image_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "total": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "ApiResponse_JobStatusResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Job status response",
                        "required": [
                            "job_id",
                            "image_id",
                            "status"
                        ],
                        "properties": {
                            "ai_model_version": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "error_message": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "finished_at": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "image_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "job_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "result_url": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "started_at": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "status": {
                                "type": "string"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "BoundingBox": {
                "type": "object",
                "description": "Bounding box for detected cell",
                "required": [
                    "class",
                    "confidence",
                    "x",
                    "y",
                    "width",
                    "height"
                ],
                "properties": {
                    "class": {
                        "type": "string"
                    },
                    "confidence": {
                        "type": "number",
                        "format": "double"
                    },
                    "height": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "width": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "x": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "y": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "CellCounts": {
                "type": "object",
                "description": "Cell counts in analysis result",
                "required": [
                    "viable",
                    "apoptosis",
                    "other"
                ],
                "properties": {
                    "apoptosis": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "other": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "viable": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "CellPercentages": {
                "type": "object",
                "description": "Cell percentages in analysis result",
                "required": [
                    "viable",
                    "apoptosis",
                    "other"
                ],
                "properties": {
                    "apoptosis": {
                        "type": "number",
                        "format": "double"
                    },
                    "other": {
                        "type": "number",
                        "format": "double"
                    },
                    "viable": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "RawDetectionData": {
                "type": "object",
                "description": "Raw detection data",
                "required": [
                    "bounding_boxes"
                ],
                "properties": {
                    "bounding_boxes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BoundingBox"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer_auth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}